feat: stabilizer (#2)
This commit adds a naive incentivized Ekubo LP staking module with fixed price 
ranges. Yin can be sent to the contract directly e.g. adding this module as a
recipient in the Allocator to receive surplus debt minted into circulation by 
the Equalizer.

The contract is intentionally limited in scope and functionality:
- In order to stake a LP token, it must correspond to the exact fee / tick spacing 
AND lower and upper price range.
- Each address can only stake one Position. If the address wants to modify its 
liquidity, then it needs to unstake, modify, and re-stake.
- Once staked, users can either (1) unstake and receive their Position NFT back; or
(2) claim accrued yin streamed to the contract.
- There is no winding down mechanism for the Stabilizer, and therefore no deadline 
for users who have staked to unstake or claim rewards.
- There is no access control.