fix(lever): add max LTV slippage check (#9)
This commit fixes a low severity finding for the Lever module whereby 
slippage on Ekubo may result in the trove's LTV after a Lever action 
exceeding what the user desired. 

The fix is to introduce a `max_ltv` argument that can be set by the user, 
and the operation would revert if the trove's LTV at the end of a Lever 
operation exceeds this value.